requirements.txt
grpcio==1.10.0
grpcio-tools==1.10.0